+++ PYME/Acquire/htsms/__init__.py
@@ -1,0 +1,8 @@
+"""
+Submodule for htsms microscope related components. This contains code that is either very specific to the htsms microscope,
+or which should be maintained in it's existing form for continuity whilst the more general platform evolves.
+
+It also includes some prototype features which are expedient to have available, but where the final implementations are
+expected to have a substantially different interface.
+
+"""
